backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.9.0
pydantic-settings==2.5.0

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.0
aiosqlite==0.20.0  # SQLite async driver for tests

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Redis
redis[hiredis]==5.1.0

# HTTP client (OpenRouter, Telegram)
httpx==0.27.0

# YouTube transcripts
youtube-transcript-api>=1.2.4
yt-dlp>=2024.1.0
curl_cffi>=0.10,<0.15

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0
httpx  # also used as test client via TestClient
